Conflicts:
	conf/distro/include/angstrom-2008-preferred-versions.inc
	conf/distro/include/preferred-xorg-versions-X11R7.5.inc
	recipes/xorg-driver/xf86-input-evdev_2.3.0.bb
	recipes/xorg-font/encodings_1.0.3.bb
	recipes/xorg-font/font-adobe-100dpi_1.0.1.bb
	recipes/xorg-font/font-adobe-75dpi_1.0.1.bb
	recipes/xorg-font/font-adobe-utopia-100dpi_1.0.2.bb
	recipes/xorg-font/font-adobe-utopia-75dpi_1.0.2.bb
	recipes/xorg-font/font-adobe-utopia-type1_1.0.2.bb
	recipes/xorg-font/font-alias_1.0.2.bb
	recipes/xorg-font/font-arabic-misc_1.0.1.bb
	recipes/xorg-font/font-bh-100dpi_1.0.1.bb
	recipes/xorg-font/font-bh-75dpi_1.0.1.bb
	recipes/xorg-font/font-bh-lucidatypewriter-100dpi_1.0.1.bb
	recipes/xorg-font/font-bh-lucidatypewriter-75dpi_1.0.1.bb
	recipes/xorg-font/font-bh-ttf_1.0.1.bb
	recipes/xorg-font/font-bh-type1_1.0.1.bb
	recipes/xorg-font/font-bitstream-100dpi_1.0.1.bb
	recipes/xorg-font/font-bitstream-75dpi_1.0.1.bb
	recipes/xorg-font/font-bitstream-speedo_1.0.1.bb
	recipes/xorg-font/font-bitstream-type1_1.0.1.bb
	recipes/xorg-font/font-cronyx-cyrillic_1.0.1.bb
	recipes/xorg-font/font-cursor-misc_1.0.1.bb
	recipes/xorg-font/font-daewoo-misc_1.0.1.bb
	recipes/xorg-font/font-dec-misc_1.0.1.bb
	recipes/xorg-font/font-ibm-type1_1.0.1.bb
	recipes/xorg-font/font-isas-misc_1.0.1.bb
	recipes/xorg-font/font-jis-misc_1.0.1.bb
	recipes/xorg-font/font-micro-misc_1.0.1.bb
	recipes/xorg-font/font-misc-cyrillic_1.0.1.bb
	recipes/xorg-font/font-misc-ethiopic_1.0.1.bb
	recipes/xorg-font/font-misc-meltho_1.0.1.bb
	recipes/xorg-font/font-mutt-misc_1.0.1.bb
	recipes/xorg-font/font-schumacher-misc_1.1.0.bb
	recipes/xorg-font/font-screen-cyrillic_1.0.2.bb
	recipes/xorg-font/font-sony-misc_1.0.1.bb
	recipes/xorg-font/font-sun-misc_1.0.1.bb
	recipes/xorg-font/font-util-native_1.1.1.bb
	recipes/xorg-font/font-util_1.1.1.bb
	recipes/xorg-font/font-winitzki-cyrillic_1.0.1.bb
	recipes/xorg-font/font-xfree86-type1_1.0.2.bb
	recipes/xorg-font/xorg-font-common.inc

+++ recipes/mtd/mtd-utils_1.2.0+git.bb
@@ -1,9 +1,7 @@
-DESCRIPTION = "Tools for managing memory technology devices."
-SECTION = "base"
-DEPENDS = "e2fsprogs-libs zlib lzo"
-HOMEPAGE = "http://www.linux-mtd.infradead.org/"
-LICENSE = "GPLv2"
-PR = "r3"
+require mtd-utils.inc
+
+DEPENDS += "e2fsprogs-libs"
+PR = "r4"
 
 ARM_INSTRUCTION_SET = "arm"
 
@@ -19,84 +17,11 @@ SRC_URI = "git://git.infradead.org/mtd-utils.git;protocol=git;tag=${TAG} \
 
 S = "${WORKDIR}/git/"
 
-EXTRA_OEMAKE = "'CC=${CC}' 'CFLAGS=${CFLAGS} -I${S}/include -DWITHOUT_XATTR'"
-
 do_configure_prepend() {
 	for i in $(find . -name "Makefile") ; do
 		sed -i -e s:lzo2:lzo:g $i
 	done
-	sed -i -e s:lzo/::g mkfs.ubifs/compr.c
+	if [ -e mkfs.ubifs/compr.c ]; then
+		sed -i -e s:lzo/::g mkfs.ubifs/compr.c
+	fi
 }
-
-do_stage () {
-	install -d ${STAGING_INCDIR}/mtd
-	for f in ${S}/include/mtd/*.h; do
-		install -m 0644 $f ${STAGING_INCDIR}/mtd/
-	done
-	for binary in ${mtd_utils}; do
-		install -m 0755 $binary ${STAGING_BINDIR}
-	done
-}
-
-mtd_utils = " docfdisk \
-	doc_loadbios \
-	flashcp \
-	flash_erase \
-	flash_eraseall \
-	flash_info \
-	flash_lock \
-	flash_otp_dump \
-	flash_otp_info \
-	flash_unlock \
-	ftl_check \
-	ftl_format \
-	jffs2dump \
-	mkfs.jffs2 \
-	mtd_debug \
-	nanddump \
-	nandtest \
-	nandwrite \
-	nftldump \
-	nftl_format \
-	recv_image \
-	rfddump \
-	rfdformat \
-	serve_image \
-	sumtool \
-	mkfs.ubifs/mkfs.ubifs \
-	ubi-utils/bin2nand \
-	ubi-utils/mkbootenv \
-	ubi-utils/nand2bin \
-	ubi-utils/pddcustomize \
-	ubi-utils/pfi2bin \
-	ubi-utils/pfiflash \
-	ubi-utils/ubigen \
-	ubi-utils/ubimirror \
-	ubi-utils/unubi \
-	ubi-utils/new-utils/ubiattach \
-	ubi-utils/new-utils/ubicrc32 \
-	ubi-utils/new-utils/ubidetach \
-	ubi-utils/new-utils/ubiformat \
-	ubi-utils/new-utils/ubimkvol \
-	ubi-utils/new-utils/ubinfo \
-	ubi-utils/new-utils/ubinize \
-	ubi-utils/new-utils/ubirmvol \
-	ubi-utils/new-utils/ubiupdatevol"
-
-
-
-do_install () {
-	install -d ${D}${bindir}
-	install -d ${D}${includedir}/mtd
-	for binary in ${mtd_utils}; do
-		install -m 0755 $binary ${D}${bindir}
-	done
-	for f in ${S}/include/mtd/*.h; do
-		install -m 0644 $f ${D}${includedir}/mtd
-	done
-}
-
-PACKAGES =+ "mkfs-jffs mkfs-jffs2"
-FILES_mkfs-jffs = "${bindir}/mkfs.jffs"
-FILES_mkfs-jffs2 = "${bindir}/mkfs.jffs2"
-
